assume we need more than one find

+++ php/ai01.php
@@ -31,14 +31,19 @@
 
   // we should check if a call is already running, we can kill that call first before
   // starting a new call
+  // this does not work, the exec call will also run the string we are looking for, so we will
+  // always detect at least one job
   exec("pgrep -f \"php/classify.R\"", $out, $return);
+  // file_put_contents("/tmp/bla.log", json_encode($out));
   if ($return == 0) {
     echo("Ok, process is running\n");
-    foreach($out as $o) {
-       if (strlen($o) > 0) {
-          echo("Kill ".$o);
-          exec("kill -9 ".$o); // would be better to kill only the current users runs
-       }
+    if (length($out) > 1) {
+      foreach($out as $o) {
+         if (strlen($o) > 0) {
+            echo("Kill ".$o);
+            exec("kill -9 ".$o); // would be better to kill only the current users runs
+         }
+      }
     }
   }
